refactor: LottoManger.test 에서 math.test로 분리
- 리뷰어 피드백 반영 (https://github.com/woowacourse/javascript-lotto/pull/370#discussion_r1965665454)
