+++ __tests__/LottoManager.test.js
@@ -1,28 +1,13 @@
 import LottoManager from '../src/Domain/Model/LottoManager.js';
 import WinningLotto from '../src/Domain/Model/WinningLotto.js';
 import Lotto from '../src/Domain/Model/Lotto.js';
-import {
-  LOTTO_DEFINITION,
-  LOTTO_PRIZE_DEFINITION,
-} from '../src/Domain/Constant/definition.js';
-import { makeNotDuplicatedRandomNumbers } from '../src/Utils/math.js';
+import { LOTTO_PRIZE_DEFINITION } from '../src/Domain/Constant/definition.js';
 
 test('로또 장수에 따라 여러 장 발행한다.', () => {
   const lottoManager = new LottoManager();
   const lottoCount = 4;
   lottoManager.makeLottoList(lottoCount);
   expect(lottoManager.getLottoList().length).toBe(4);
-});
-
-test('로또 1장당 범위내에서 중복되지 않는 랜덤한 번호 6개를 만든다.', () => {
-  const lottoNumbers = makeNotDuplicatedRandomNumbers(
-    LOTTO_DEFINITION.NUMBER_COUNTS,
-    {
-      min: LOTTO_DEFINITION.MIN_NUMBER,
-      max: LOTTO_DEFINITION.MAX_NUMBER,
-    }
-  );
-  expect(lottoNumbers.length).toBe(LOTTO_DEFINITION.NUMBER_COUNTS);
 });
 
 test('당첨 내역을 반환한다.', () => {
